requirements.txt
cvlib
Flask
opencv_python
Pillow
